+++ modules/tinymce/src/core/main/ts/fmt/FormatUtils.ts
@@ -216,6 +216,10 @@ const isSelectorFormat = (format: ApplyFormat): format is SelectorFormat =>
 const isInlineFormat = (format: ApplyFormat): format is InlineFormat =>
   Obj.hasNonNullableKey(format as any, 'inline');
 
+const hasBlockChildren = (dom: DOMUtils, elm: Node) => Arr.exists(elm.childNodes, dom.isBlock);
+
+const isChildText = (nodes: Node[], node: Node) => NodeType.isText(node) && Arr.contains(nodes, node);
+
 export {
   isNode,
   isInlineBlock,
@@ -235,5 +239,7 @@ export {
   areSimilarFormats,
   isSelectorFormat,
   isInlineFormat,
-  isBlockFormat
+  isBlockFormat,
+  hasBlockChildren,
+  isChildText
 };